+++ src/imio/dashboard/browser/overrides.py
@@ -81,7 +81,8 @@ class IDDocumentGenerationView(DocumentGenerationView):
                     if isinstance(v, list) and len(v) == 1:
                         v = v[0]
                     self.request.form[k] = v
-            brains = faceted_query.query(batch=False)
+            import ipdb; ipdb.set_trace()
+            brains = faceted_query.query(batch=False, sort=True)
             uids = [brain.UID for brain in brains]
         else:
             uids = uids.split(',')